pyproject.toml
[project]
name = "mujoco-mcp-server"
version = "0.2.0"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.26",
    "mujoco>=2.3.1",
    "numpy>=1.24",
    "Pillow>=10.0",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "packaging>=21.0",
]

[project.scripts]
mujoco-mcp = "mujoco_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mujoco_mcp"]
